Extract types.ts
