+++ hub-browser-client/src/api/llamafile_api.ts
@@ -1,9 +1,4 @@
-export interface ILlamafile {
-  name: string;
-  url: string;
-  downloaded: boolean;
-  running: boolean;
-}
+import { ILlamafile } from "../types";
 
 const host = "http://localhost:8001";
 const llamafileApi = `${host}/api/llamafile`;
