+++ hub-browser-client/src/types.ts
@@ -1,0 +1,6 @@
+export interface ILlamafile {
+  name: string;
+  url: string;
+  downloaded: boolean;
+  running: boolean;
+}
